Conditional version check for resources
Currently there is an envvar for the GRAFANA_VERSION.
But even if you set it, the lib still calls out to the health check API.
In the case of AWS-managed Grafana that API is disabled.
It's best that if the GRAFANA_VERSION is set we simply skip the API
request.